deployment/requirements.txt
streamlit
pandas
numpy
scikit.learn==1.3.0
tensorflow==2.18.0
tensorflow_hub
keras==3.5.0
pillow
opencv-python